+++ examples/lib/main.dart
@@ -9,8 +9,10 @@ import 'package:examples/stories/bridge_libraries/forge2d/flame_forge2d.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/constant_volume_joint.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/distance_joint.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/friction_joint.dart';
+import 'package:examples/stories/bridge_libraries/forge2d/joints/gear_joint.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/motor_joint.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/mouse_joint.dart';
+import 'package:examples/stories/bridge_libraries/forge2d/joints/prismatic_joint.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/pulley_joint.dart';
 import 'package:examples/stories/bridge_libraries/forge2d/joints/revolute_joint.dart';
 import 'package:examples/stories/camera_and_viewport/camera_and_viewport.dart';
@@ -40,9 +42,11 @@ void main() {
     'constant_volume_joint': ConstantVolumeJointExample.new,
     'distance_joint': DistanceJointExample.new,
     'friction_joint': FrictionJointExample.new,
+    'gear_joint': GearJointExample.new,
     'motor_joint': MotorJointExample.new,
     'mouse_joint': MouseJointExample.new,
     'pulley_joint': PulleyJointExample.new,
+    'prismatic_joint': PrismaticJointExample.new,
     'revolute_joint': RevoluteJointExample.new,
   };
   final game = routes[page]?.call();